requirements.txt
fastmcp
git
jinja2
openai
pandas
pyaml
regex
requests